+++ platforms/reference/src/SimTKReference/ReferenceRigidShakeAlgorithm.cpp
@@ -359,8 +359,8 @@ int ReferenceRigidShakeAlgorithm::apply( int numberOfAtoms, RealOpenMM** atomCoo
               r[j][2] *= invLength;
           }
           Array2D<double> matrix(size, size);
-          for (unsigned int j = 0; j < size; j++) {
-              for (unsigned int k = 0; k < size; k++) {
+          for (int j = 0; j < (int)size; j++) {
+              for (int k = 0; k < (int)size; k++) {
                   double dot;
                   int atomj0 = _atomIndices[cluster[j]][0];
                   int atomj1 = _atomIndices[cluster[j]][1];
@@ -390,20 +390,20 @@ int ReferenceRigidShakeAlgorithm::apply( int numberOfAtoms, RealOpenMM** atomCoo
           svd.getV(v);
           svd.getSingularValues(w);
           double singularValueCutoff = 0.01*w[0];
-          for (unsigned int j = 0; j < size; j++)
+          for (int j = 0; j < (int)size; j++)
               w[j] = (w[j] < singularValueCutoff ? 0.0 : 1.0/w[j]);
-          for (unsigned int j = 0; j < size; j++) {
-              for (unsigned int k = 0; k < size; k++) {
+          for (int j = 0; j < (int)size; j++) {
+              for (int k = 0; k < (int)size; k++) {
                   matrix[j][k] = 0.0;
-                  for (unsigned int m = 0; m < size; m++)
+                  for (int m = 0; m < (int)size; m++)
                       matrix[j][k] += v[j][m]*w[m]*u[k][m];
               }
           }
 
           // Record the inverted matrix.
 
-          for (unsigned int j = 0; j < size; j++)
-              for (unsigned int k = 0; k < size; k++)
+          for (int j = 0; j < (int)size; j++)
+              for (int k = 0; k < (int)size; k++)
                   _matrices[i][j][k] = (RealOpenMM)matrix[j][k]*_distance[cluster[k]]/_distance[cluster[j]];
       }
    }
